requirements.txt
ansible==7.6.0
cryptography==41.0.1
dnspython==2.4.2
jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
netaddr==0.8.0
pbr==5.11.1
ruamel.yaml==0.17.31
ruamel.yaml.clib>=0.2.7
